requirements.txt
httpx==0.27.0
pywebio==1.8.3
peewee==3.17.5
contextvars==2.4
pydantic==2.7.3
fastapi~=0.111.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1